refactor: extract server-auth.js from server.js (PR-15)
Move all authentication logic into dedicated server-auth module:
PIN auth, multi-user login, OTP flow, invite registration,
admin setup, logout, password recovery, and rate limiting.

server.js delegates auth routes via auth.handleRequest() and
accesses shared helpers (getMultiUserFromReq, isRequestAuthed,
parseCookies) through the returned auth context object.

server.js: 3778 -> 3086 lines. server-auth.js: 737 lines.
